[NO-TASK] Fixed build error
