+++ src/subdomains/user/api/controllers/user.controller.ts
@@ -4,7 +4,10 @@ import { FilesInterceptor } from '@nestjs/platform-express';
 import { ApiBearerAuth, ApiCreatedResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger';
 import { GetJwt } from 'src/shared/auth/get-jwt.decorator';
 import { JwtPayload } from 'src/shared/auth/jwt-payload.interface';
+import { CountryDtoMapper } from 'src/subdomains/master-data/country/dto/country-dto.mapper';
+import { CountryDto } from 'src/subdomains/master-data/country/dto/country.dto';
 import { KycService } from '../../services/kyc.service';
+import { MandatorService } from '../../services/mandator.service';
 import { UserService } from '../../services/user.service';
 import { KycDataDto, SettingsDto } from '../dto/user-in.dto';
 import { UserInfoDto } from '../dto/user-out.dto';
@@ -12,7 +15,11 @@ import { UserInfoDto } from '../dto/user-out.dto';
 @ApiTags('User')
 @Controller('user')
 export class UserController {
-  constructor(private readonly userService: UserService, private readonly kycService: KycService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly mandatorService: MandatorService,
+    private readonly kycService: KycService,
+  ) {}
 
   // --- USER --- //
   @Get()
@@ -29,6 +36,14 @@ export class UserController {
   @ApiOkResponse({ type: UserInfoDto })
   async updateUser(@GetJwt() { mandator, user }: JwtPayload, settings: SettingsDto): Promise<UserInfoDto> {
     return this.userService.updateSettings(mandator, user, settings);
+  }
+
+  @Get('countries')
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard())
+  @ApiOkResponse({ type: CountryDto, isArray: true })
+  async getCountries(@GetJwt() { mandator }: JwtPayload): Promise<CountryDto[]> {
+    return this.mandatorService.getAllowedCountries(mandator).then(CountryDtoMapper.entitiesToDto);
   }
 
   // --- KYC --- //
