+++ src/subdomains/user/entities/kyc.enum.ts
@@ -1,0 +1,13 @@
+export enum KycStepName {
+  USER_DATA = 'UserData',
+  FILE_UPLOAD = 'FileUpload',
+  CHATBOT = 'Chatbot',
+  ONLINE_ID = 'OnlineId',
+  VIDEO_ID = 'VideoId',
+}
+
+export enum KycStepStatus {
+  IN_PROGRESS = 'InProgress',
+  FAILED = 'Failed',
+  COMPLETED = 'Completed',
+}
