+++ src/subdomains/user/entities/user.enum.ts
@@ -1,0 +1,12 @@
+export enum KycStatus {
+  NA = 'NA',
+  IN_PROGRESS = 'InProgress',
+  COMPLETED = 'Completed',
+  PAUSED = 'Paused',
+}
+
+export enum AccountType {
+  PERSONAL = 'Personal',
+  BUSINESS = 'Business',
+  SOLE_PROPRIETORSHIP = 'SoleProprietorship',
+}
